# Conflicts:
#	src/physbo/misc/_set_config.py

+++ docs/make_docs.sh
@@ -1,11 +1,19 @@
 #!/bin/bash
 
-if ! which sphinx-build > /dev/null 2>&1; then
+if ! which sphinx-build >/dev/null 2>&1; then
   echo "ERROR: Sphinx is not installed"
   exit 1
 fi
 
-ROOT_DIR=$(cd $(dirname $0)/..; pwd)
+LANGS="ja en"
+if [[ $# -gt 0 ]]; then
+  LANGS="$*"
+fi
+
+ROOT_DIR=$(
+  cd $(dirname $0)/..
+  pwd
+)
 
 DOCS_DIR="$ROOT_DIR/docs/sphinx/manual"
 DOCS_OUT="$ROOT_DIR/docs/built"
@@ -14,7 +22,12 @@ DOCS_OUT="$ROOT_DIR/docs/built"
 rm -rf "$DOCS_OUT"
 mkdir -p "$DOCS_OUT"
 
-for lang in ja en; do
+for lang in ${LANGS}; do
+  if [[ $lang != "ja" && $lang != "en" ]]; then
+    echo "ERROR: Unsupported language: $lang"
+    exit 1
+  fi
+
   # Generate API documentation
   rm -rf "$DOCS_DIR/$lang/source/api"
   sphinx-apidoc -f -T -d 3 -e -o "$DOCS_DIR/$lang/source/api" "$ROOT_DIR/src/physbo"
